Audio fix
Thanks Layer812 I now understand the relationships.
I was able to set your preferred settings with the wrap, and still get the volume boost on the piezo.
I have it set to * 2.8f, 2-4 seems to be a decent range. however playing at 100 volume on some games can be too much.
like blade buster hb is a bit much at 100, but like smb 100 is great.